Improves test coverage
- Added Mockery dependeces
- Organizing tests
