+++ composer.json
@@ -22,7 +22,8 @@
   "require-dev": {
     "phpunit/phpunit": "^5.7 || ^6.0",
     "orchestra/testbench": "^3.2",
-    "squizlabs/php_codesniffer": "^2.3"
+    "squizlabs/php_codesniffer": "^2.3",
+    "mockery/mockery": "~0.9.4"
   },
   "autoload": {
     "psr-4": {
